refactor: put select and confirmation into each document type
