+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/brp/select/SelectBrpScreen.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.confirmation
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp.select
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -7,17 +7,17 @@ import uk.gov.android.ui.theme.m3.GdsTheme
 import uk.gov.onelogin.criorchestrator.libraries.composeutils.LightDarkBothLocalesPreview
 
 @Composable
-internal fun ConfirmDocumentScreen(modifier: Modifier = Modifier) {
+internal fun SelectBrpScreen(modifier: Modifier = Modifier) {
     Text(
-        text = "DCMAW-8798 | Android | Document Selection |  Photo ID confirmation screen",
+        text = "DCMAW-10690 | Android | Document Selection | Do you have a BRP screen",
         modifier = modifier,
     )
 }
 
 @LightDarkBothLocalesPreview
 @Composable
-internal fun PreviewConfirmDocumentScreen() {
+internal fun PreviewSelectBrpScreen() {
     GdsTheme {
-        ConfirmDocumentScreen()
+        SelectBrpScreen()
     }
 }
