+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/passport/confirmation/ConfirmDocumentScreen.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.confirmation
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -7,17 +7,17 @@ import uk.gov.android.ui.theme.m3.GdsTheme
 import uk.gov.onelogin.criorchestrator.libraries.composeutils.LightDarkBothLocalesPreview
 
 @Composable
-internal fun SelectBrpScreen(modifier: Modifier = Modifier) {
+internal fun ConfirmDocumentScreen(modifier: Modifier = Modifier) {
     Text(
-        text = "DCMAW-10690 | Android | Document Selection | Do you have a BRP screen",
+        text = "DCMAW-8798 | Android | Document Selection |  Photo ID confirmation screen",
         modifier = modifier,
     )
 }
 
 @LightDarkBothLocalesPreview
 @Composable
-internal fun PreviewSelectBrpScreen() {
+internal fun PreviewConfirmDocumentScreen() {
     GdsTheme {
-        SelectBrpScreen()
+        ConfirmDocumentScreen()
     }
 }
