+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/passport/select/SelectPassportViewModelModule.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.select
 
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewmodel.viewModelFactory
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocumentAnalytics
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocAnalytics
 import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Named
 
@@ -17,7 +17,7 @@ object SelectPassportViewModelModule {
 
     @Provides
     @Named(FACTORY_NAME)
-    fun provideFactory(analytics: SelectDocumentAnalytics): ViewModelProvider.Factory =
+    fun provideFactory(analytics: SelectDocAnalytics): ViewModelProvider.Factory =
         viewModelFactory {
             initializer {
                 SelectPassportViewModel(analytics)
