+++ features/select-doc/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/analytics/SelectDocumentAnalyticsTest.kt
@@ -24,7 +24,7 @@ class SelectDocumentAnalyticsTest {
     private val context: Context = ApplicationProvider.getApplicationContext()
 
     private val analytics by lazy {
-        SelectDocumentAnalytics(
+        SelectDocAnalytics(
             resourceProvider = AndroidResourceProvider(context = context),
             analyticsLogger = analyticsLogger,
         )
