+++ features/select-doc/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/passport/SelectPassportScreenAnalyticsTest.kt
@@ -19,8 +19,10 @@ import uk.gov.logging.api.v3dot1.logger.asLegacyEvent
 import uk.gov.logging.api.v3dot1.model.TrackEvent
 import uk.gov.logging.api.v3dot1.model.ViewEvent
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.R
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocumentAnalytics
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocAnalytics
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocumentScreenId
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.select.SelectPassportScreen
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.select.SelectPassportViewModel
 import uk.gov.onelogin.criorchestrator.libraries.analytics.resources.AndroidResourceProvider
 import uk.gov.onelogin.criorchestrator.libraries.testing.MainStandardDispatcherRule
 import uk.gov.onelogin.criorchestrator.libraries.testing.ReportingAnalyticsLoggerRule
@@ -44,7 +46,7 @@ class SelectPassportScreenAnalyticsTest {
     private val image = hasContentDescription(context.getString(R.string.selectdocument_passport_imagedescription))
 
     private val analytics =
-        SelectDocumentAnalytics(
+        SelectDocAnalytics(
             resourceProvider =
                 AndroidResourceProvider(
                     context = context,
@@ -64,7 +66,7 @@ class SelectPassportScreenAnalyticsTest {
                 .Screen(
                     id = SelectDocumentScreenId.SelectPassport.rawId,
                     name = context.getString(R.string.selectdocument_passport_title),
-                    params = SelectDocumentAnalytics.requiredParameters,
+                    params = SelectDocAnalytics.requiredParameters,
                 ).asLegacyEvent()
         composeTestRule.setSelectPassportScreenContent()
         assertContains(analyticsLogger.loggedEvents, expectedEvent)
@@ -88,7 +90,7 @@ class SelectPassportScreenAnalyticsTest {
             TrackEvent
                 .Form(
                     text = context.getString(R.string.selectdocument_passport_continuebutton),
-                    params = SelectDocumentAnalytics.requiredParameters,
+                    params = SelectDocAnalytics.requiredParameters,
                     response = context.getString(R.string.selectdocument_passport_selection_yes),
                 ).asLegacyEvent()
         assertContains(analyticsLogger.loggedEvents, expectedEvent)
@@ -108,7 +110,7 @@ class SelectPassportScreenAnalyticsTest {
             TrackEvent
                 .Button(
                     text = context.getString(R.string.selectdocument_passport_readmore_button),
-                    params = SelectDocumentAnalytics.requiredParameters,
+                    params = SelectDocAnalytics.requiredParameters,
                 ).asLegacyEvent()
         assertContains(analyticsLogger.loggedEvents, expectedEvent)
     }
