+++ features/select-doc/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/passport/SelectPassportViewModelTest.kt
@@ -10,18 +10,20 @@ import org.junit.jupiter.params.provider.ValueSource
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.R
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocumentAnalytics
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocAnalytics
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocumentScreenId
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.select.SelectPassportAction
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.select.SelectPassportViewModel
 import uk.gov.onelogin.criorchestrator.libraries.testing.MainDispatcherExtension
 
 @ExtendWith(MainDispatcherExtension::class)
 class SelectPassportViewModelTest {
-    private val analyticsLogger = mock<SelectDocumentAnalytics>()
+    private val analyticsLogger = mock<SelectDocAnalytics>()
 
     private val viewModel by lazy {
         SelectPassportViewModel(
             analytics =
-            analyticsLogger,
+                analyticsLogger,
         )
     }
